requirements-file.txt
biopython==1.81
numpy==1.24.3
matplotlib==3.7.1
